coding-mbti/backend/requirements.txt
asgiref==3.2.10
astroid==2.4.2
attrs==20.2.0
certifi==2020.6.20
chardet==3.0.4
coverage==5.3
coveralls==2.1.2
Django==3.1.2
django-cors-headers==3.5.0
django-extensions==3.0.9
djangorestframework==3.12.2
docopt==0.6.2
gunicorn==20.0.4
idna==2.10
importlib-metadata==2.0.0
iniconfig==1.1.1
isort==5.6.4
joblib==0.17.0
lazy-object-proxy==1.4.3
mccabe==0.6.1
numpy==1.19.4
packaging==20.4
Pillow==8.0.1
pluggy==0.13.1
psycopg2==2.8.6
py==1.9.0
pygraphviz==1.6
pylint==2.6.0
pylint-django==2.3.0
pylint-plugin-utils==0.6
pyparsing==2.4.7
pytest==6.1.2
pytz==2020.1
requests==2.24.0
scikit-learn==0.23.2
scipy==1.5.4
six==1.15.0
sqlparse==0.4.1
threadpoolctl==2.1.0
toml==0.10.1
typed-ast==1.4.1
urllib3==1.25.11
wrapt==1.12.1
zipp==3.4.0
